+++ crates/ide_completion/src/patterns.rs
@@ -13,10 +13,10 @@ pub(crate) fn determine_location(
     let node = token.parent()?;
     let parent = node.parent()?;
 
-    if let Some(expr) = ast::FieldExpr::cast(node.clone()) {
-        Some(ImmediateLocation::FieldAccess { expr })
-    } else if let Some(expr) = ast::FieldExpr::cast(parent.clone()) {
-        Some(ImmediateLocation::FieldAccess { expr })
+    if let Some(expression) = ast::FieldExpression::cast(node.clone()) {
+        Some(ImmediateLocation::FieldAccess { expression })
+    } else if let Some(expression) = ast::FieldExpression::cast(parent.clone()) {
+        Some(ImmediateLocation::FieldAccess { expression })
     } else if node.kind() == SyntaxKind::SourceFile {
         Some(ImmediateLocation::ItemList)
     } else if node.kind() == SyntaxKind::Import || parent.kind() == SyntaxKind::Import {